requirements.txt
pip install -r requirements.txtasgiref==3.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-filter==23.4
django-import-export==3.3.3
django-simple-history==3.4.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
future==0.18.3
importlib-metadata==6.8.0
Markdown==3.5.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pytz==2023.3.post1
PyYAML==6.0.1
serializers==0.2.4
sqlparse==0.4.4
tablib==3.5.0
typing-extensions==4.8.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.17.0